Add support for setting the version in debian.changelog
For Debian builds the primary location of the package version number is the
changelog file. In some cases where set_version is used in combination with
tar_scm's versionformat parameter it might be required to amend the
Debian changelog as well. Later during the build debtransform will assemble
the correct format of the sources.

Signed-off-by: Jan Blunck <jblunck@infradead.org>