Avoid half-converting Debian native pkgs to non-native pkgs
Newer versions of dpkg-buildpackages are more strict in checking that
the Debian package format is consistent.
Always appending a -0 to the version causes a native package, which
must not have a - in the version, to fail to build.
Append -0 conditionally on a simple check: if the original version
has a - character then it's non-native and it's safe to add -0.

In order to do this check a new version detection static method is
added, to parse the .dsc file for the Version tag.