fix suffixes to begin with a dot
Without this fix, if a package ends with one of the strings in the suffix
(like e.g. pristine-tar) set_version tries to unpack it. In this example
it finds a directory which ends with "tar" which is the leftover from the
tar_scm service and the attempt of unpacking results in an exception.

```
  File "/usr/lib64/python3.8/gzip.py", line 173, in __init__
    fileobj = self.myfileobj = builtins.open(filename, mode or 'rb')
IsADirectoryError: [Errno 21] Is a directory: 'pristine-tar'
```

This fix contains 2 steps:

* check only files with ".<suffix>" at the end instead of "<suffix>"
* check matching files if they are regular files before trying to unpack